feat: wip - add compressed memory cache

+++ orebfuscator-common/src/main/java/net/imprex/orebfuscator/config/AdvancedConfig.java
@@ -12,7 +12,7 @@ public interface AdvancedConfig {
 
 	boolean hasObfuscationTimeout();
 
-	int obfuscationTimeout();
+	long obfuscationTimeout();
 
 	int proximityHiderThreads();
